pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "san-zone-designer"
version = "1.1.0"
description = "SAN Zone Designer - Cisco/Brocade SAN Config Generator"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    { name = "Konrad Łatkowski" },
]
dependencies = [
    "typer>=0.9",
    "rich>=13",
    "InquirerPy>=0.3",
    "pyyaml>=6",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
]
web = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "python-multipart>=0.0.6",
    "bcrypt>=4.0",
]

[project.scripts]
san-zone-designer = "san_zone_designer.cli:main"

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
